feat: Implement vector processing algorithms with NE and NS data handling
- Added NeData and NsData data classes for structured representation of grid and edge data.
- Implemented write_ne and write_ns functions to output NE and NS data to files.
- Developed get_ne and get_ns functions to read and parse NE and NS data from files.
- Introduced geometric functions for point-in-polygon checks and intersection tests with GeoJSON features.
- Implemented coordinate transformation functions using caching for efficiency.
- Added grid finding functions to associate points and features with grid IDs.
- Enhanced logging for better traceability during data processing.
